feat: Add astro-icon npm dependency and DiscountProduct model
This commit adds the "astro-icon" npm dependency to the project, which will be used for displaying icons. It also introduces a new model called "DiscountProduct" to the app/Models directory. The DiscountProduct model extends the Eloquent Model class and uses the HasFactory trait. This model will be used to represent discounted products in the application.

Note: This commit message follows the established convention of using a prefix (feat) to indicate a new feature or functionality.